// src/io/resolver/types.ts
/**
 * Copyright (C) 2023 Expedia, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Key, Value } from '../../model/Types';

export interface ResolverOptions {
  baseUrl?: string;
  timeout?: number;
  maxDepth?: number;
  headers?: Record<string, string>;
  cache?: boolean;
}

export interface ResolvedRef {
  url: string;
  fragment: string;
  content: Record<Key, Value>;
}